# openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "SplitWiseMoney API",
                 "description":  "SplitWiseMoney - Expense sharing web application API documentation",
                 "version":  "1.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Generated server url"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/settlements/{settlementId}/settle":  {
                                                                 "put":  {
                                                                             "tags":  [
                                                                                          "Settlements"
                                                                                      ],
                                                                             "summary":  "Mark a pending settlement payment as SETTLED",
                                                                             "operationId":  "markAsSettled",
                                                                             "parameters":  [
                                                                                                {
                                                                                                    "name":  "settlementId",
                                                                                                    "in":  "path",
                                                                                                    "required":  true,
                                                                                                    "schema":  {
                                                                                                                   "type":  "integer",
                                                                                                                   "format":  "int64"
                                                                                                               }
                                                                                                }
                                                                                            ],
                                                                             "responses":  {
                                                                                               "200":  {
                                                                                                           "description":  "OK",
                                                                                                           "content":  {
                                                                                                                           "*/*":  {
                                                                                                                                       "schema":  {
                                                                                                                                                      "$ref":  "#/components/schemas/SettlementResponse"
                                                                                                                                                  }
                                                                                                                                   }
                                                                                                                       }
                                                                                                       }
                                                                                           },
                                                                             "security":  [
                                                                                              {
                                                                                                  "bearerAuth":  [

                                                                                                                 ]
                                                                                              }
                                                                                          ]
                                                                         }
                                                             },
...
                  "/api/settlements/group/{groupId}":  {
                                                           "get":  {
                                                                       "tags":  [
                                                                                    "Settlements"
                                                                                ],
                                                                       "summary":  "List all recorded settlement transactions in a group",
                                                                       "operationId":  "getGroupSettlements",
                                                                       "parameters":  [
                                                                                          {
                                                                                              "name":  "groupId",
                                                                                              "in":  "path",
                                                                                              "required":  true,
                                                                                              "schema":  {
                                                                                                             "type":  "integer",
                                                                                                             "format":  "int64"
                                                                                                         }
                                                                                          }
                                                                                      ],
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "OK",
                                                                                                     "content":  {
                                                                                                                     "*/*":  {
                                                                                                                                 "schema":  {
                                                                                                                                                "type":  "array",
                                                                                                                                                "items":  {
                                                                                                                                                              "$ref":  "#/components/schemas/SettlementResponse"
                                                                                                                                                          }
                                                                                                                                            }
                                                                                                                             }
                                                                                                                 }
                                                                                                 }
                                                                                     },
                                                                       "security":  [
                                                                                        {
                                                                                            "bearerAuth":  [

                                                                                                           ]
                                                                                        }
                                                                                    ]
                                                                   },
                                                           "post":  {
                                                                        "tags":  [
                                                                                     "Settlements"
                                                                                 ],
                                                                        "summary":  "Record a settlement payment between group members",
                                                                        "operationId":  "createSettlement",
                                                                        "parameters":  [
                                                                                           {
                                                                                               "name":  "groupId",
                                                                                               "in":  "path",
                                                                                               "required":  true,
                                                                                               "schema":  {
                                                                                                              "type":  "integer",
                                                                                                              "format":  "int64"
                                                                                                          }
                                                                                           }
                                                                                       ],
                                                                        "requestBody":  {
                                                                                            "content":  {
                                                                                                            "application/json":  {
                                                                                                                                     "schema":  {
                                                                                                                                                    "$ref":  "#/components/schemas/SettlementRequest"
                                                                                                                                                }
                                                                                                                                 }
                                                                                                        },
                                                                                            "required":  true
                                                                                        },
                                                                        "responses":  {
                                                                                          "200":  {
                                                                                                      "description":  "OK",
                                                                                                      "content":  {
                                                                                                                      "*/*":  {
                                                                                                                                  "schema":  {
                                                                                                                                                 "$ref":  "#/components/schemas/SettlementResponse"
                                                                                                                                             }
                                                                                                                              }
                                                                                                                  }
                                                                                                  }
                                                                                      },
                                                                        "security":  [
                                                                                         {
                                                                                             "bearerAuth":  [

                                                                                                            ]
                                                                                         }
                                                                                     ]
                                                                    }
                                                       },
...
                  "/api/settlements/group/{groupId}/owed":  {
                                                                "get":  {
                                                                            "tags":  [
                                                                                         "Settlements"
                                                                                     ],
                                                                            "summary":  "Calculate simplified owed debts for members of a group (minimum transactions)",
                                                                            "operationId":  "getOwedSettlements",
                                                                            "parameters":  [
                                                                                               {
                                                                                                   "name":  "groupId",
                                                                                                   "in":  "path",
                                                                                                   "required":  true,
                                                                                                   "schema":  {
                                                                                                                  "type":  "integer",
                                                                                                                  "format":  "int64"
                                                                                                              }
                                                                                               }
                                                                                           ],
                                                                            "responses":  {
                                                                                              "200":  {
                                                                                                          "description":  "OK",
                                                                                                          "content":  {
                                                                                                                          "*/*":  {
                                                                                                                                      "schema":  {
                                                                                                                                                     "type":  "array",
                                                                                                                                                     "items":  {
                                                                                                                                                                   "$ref":  "#/components/schemas/SettlementResponse"
                                                                                                                                                               }
                                                                                                                                                 }
                                                                                                                                  }
                                                                                                                      }
                                                                                                      }
                                                                                          },
                                                                            "security":  [
                                                                                             {
                                                                                                 "bearerAuth":  [

                                                                                                                ]
                                                                                             }
                                                                                         ]
                                                                        }
                                                            },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "SettlementResponse":  {
                                                                  "type":  "object",
                                                                  "properties":  {
                                                                                     "id":  {
                                                                                                "type":  "integer",
                                                                                                "format":  "int64"
                                                                                            },
                                                                                     "fromUserId":  {
                                                                                                        "type":  "integer",
                                                                                                        "format":  "int64"
                                                                                                    },
                                                                                     "fromUserName":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                     "toUserId":  {
                                                                                                      "type":  "integer",
                                                                                                      "format":  "int64"
                                                                                                  },
                                                                                     "toUserName":  {
                                                                                                        "type":  "string"
                                                                                                    },
                                                                                     "amount":  {
                                                                                                    "type":  "number"
                                                                                                },
                                                                                     "status":  {
                                                                                                    "type":  "string"
                                                                                                },
                                                                                     "createdAt":  {
                                                                                                       "type":  "string",
                                                                                                       "format":  "date-time"
                                                                                                   }
                                                                                 }
                                                              },
...
                                       "SettlementRequest":  {
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "fromUserId":  {
                                                                                                       "type":  "integer",
                                                                                                       "format":  "int64"
                                                                                                   },
                                                                                    "toUserId":  {
                                                                                                     "type":  "integer",
                                                                                                     "format":  "int64"
                                                                                                 },
                                                                                    "amount":  {
                                                                                                   "type":  "number",
                                                                                                   "minimum":  0.01
                                                                                               },
                                                                                    "status":  {
                                                                                                   "type":  "string"
                                                                                               }
                                                                                },
                                                                 "required":  [
                                                                                  "amount",
                                                                                  "fromUserId",
                                                                                  "status",
                                                                                  "toUserId"
                                                                              ]
                                                             },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "name":  "bearerAuth",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}